refactor: TrickRules und BiddingRules in Rule Package verschoben
