+++ src/main/java/com/aau/wizard/service/impl/RoundServiceImpl.java
@@ -8,9 +8,9 @@ import com.aau.wizard.model.Player;
 import com.aau.wizard.model.enums.CardSuit;
 import com.aau.wizard.model.enums.GameStatus;
 import com.aau.wizard.service.interfaces.GameService;
-import com.aau.wizard.util.BiddingRules;
+import com.aau.wizard.rules.BiddingRules;
 import com.aau.wizard.util.Pair;
-import com.aau.wizard.util.TrickRules;
+import com.aau.wizard.rules.TrickRules;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
 import java.util.ArrayList;
 import java.util.HashMap;
